rename EditingSurface -> EditorSurface
