+++ src/components/EditorSurface/Canvas.test.js
