+++ src/components/EditorSurface/CanvasSection.story.js
