+++ src/components/EditorSurface/CanvasSection.test.js
