+++ src/components/SeamlessInput/SeamlessInput.js
@@ -1,5 +1,5 @@
 import React from "react";
-import withSeamlessness from "../EditingSurface/withSeamlessness";
+import withSeamlessness from "../EditorSurface/withSeamlessness";
 import withChangeHandler from "components/Forms/withChangeHandler";
 import { flow } from "lodash";
