+++ src/components/SeamlessTextArea/SeamlessTextArea.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import AutoResizeTextArea from "react-textarea-autosize";
-import withSeamlessness from "../EditingSurface/withSeamlessness";
+import withSeamlessness from "../EditorSurface/withSeamlessness";
 import withChangeHandler from "components/Forms/withChangeHandler";
 import { flow } from "lodash";
